requirements.txt
@torch>=2.0.1
diffusers==0.33.1
transformers
numpy
Pillow
comfyui
